backend/requirements.txt
# Dressly Backend Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
pymongo==4.10.1

# HTTP Client
httpx==0.27.2

# AI/ML
google-generativeai==0.8.3

# Environment Variables
python-dotenv==1.0.1

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# CORS (included with fastapi)
